+++ app/src/main/kotlin/com/example/seremtinameno/datamanager/core/AndroidApplication.kt
@@ -1,10 +1,14 @@
 package com.example.seremtinameno.datamanager.core
 
 import android.app.Application
+import android.content.ContextWrapper
+import com.crashlytics.android.Crashlytics
 import com.example.seremtinameno.datamanager.core.di.ApplicationComponent
 import com.example.seremtinameno.datamanager.core.di.ApplicationModule
 import com.example.seremtinameno.datamanager.core.di.DaggerApplicationComponent
+import com.pixplicity.easyprefs.library.Prefs
 import com.squareup.leakcanary.LeakCanary
+import io.fabric.sdk.android.Fabric
 import org.koin.android.ext.android.startKoin
 
 
@@ -19,8 +23,11 @@ class AndroidApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
-        this.injectMembers()
-        this.initializeLeakDetection()
+        injectMembers()
+        initializeLeakDetection()
+
+        initPrefs()
+        initCrashlytics()
     }
 
     private fun injectMembers() = appComponent.inject(this)
@@ -29,4 +36,20 @@ class AndroidApplication : Application() {
 //        if (BuildConfig.DEBUG)
             LeakCanary.install(this)
     }
+
+    private fun initPrefs() {
+        Prefs.Builder()
+                .setContext(this)
+                .setMode(ContextWrapper.MODE_PRIVATE)
+                .setPrefsName(packageName)
+                .setUseDefaultSharedPreference(true)
+                .build()
+    }
+
+
+    private fun initCrashlytics() {
+        Fabric.with(this, Crashlytics())
+    }
+
+
 }
