+++ app/src/main/res/drawable-v24/rounded_button_color_border.xml
@@ -2,8 +2,8 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
 
-    <!--<solid android:color="@color/green"/>-->
+    <!--<solid android:color="@color/white"/>-->
     <corners android:radius="20dp"/>
-    <stroke android:width="10px" android:color="@color/white"/>
+    <stroke android:color="@color/white" android:width="10px"/>
 
 </shape>
