+++ app/src/main/res/drawable/app_background_selector.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape>
+            <gradient
+                android:angle="90"
+                android:startColor="@color/black"
+                android:endColor="#27304D"
+                android:type="linear"
+                />
+        </shape>
+    </item>
+</selector>
+    <!--171622-->
