+++ app/src/main/res/layout/activity_set_data_limit.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".features.datausage.SetDataLimitActivity"
-    android:background="@color/dark_blue"
+    android:background="@drawable/app_background_selector"
     >
 
     <LinearLayout
@@ -80,8 +79,8 @@
                 android:id="@+id/cancel"
                 android:layout_width="100dp"
                 android:layout_height="35dp"
-                android:background="@drawable/rounded_button"
-                android:textColor="@color/dark_blue"
+                android:background="@drawable/rounded_button_color_border"
+                android:textColor="@color/white"
                 android:text="@string/cancel"
                 android:layout_marginEnd="12dp"
                 />
@@ -90,8 +89,8 @@
                 android:id="@+id/confirm"
                 android:layout_width="100dp"
                 android:layout_height="35dp"
-                android:background="@drawable/rounded_button"
-                android:textColor="@color/dark_blue"
+                android:background="@drawable/rounded_button_color_border"
+                android:textColor="@color/white"
                 android:text="@string/confirm"
                 android:layout_marginStart="12dp"
                 />
