+++ app/src/main/res/layout/fragment_daily.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.github.ksoichiro.android.observablescrollview.ObservableScrollView android:id="@+id/scrollView"
+<com.github.ksoichiro.android.observablescrollview.ObservableScrollView
+    android:id="@+id/scrollView"
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:scrollbars="none"
-    android:background="@color/dark_blue"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <LinearLayout
@@ -32,11 +32,10 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:text="May 24"
-                android:textColor="@color/dark_blue"
+                android:textColor="@color/white"
                 android:textSize="25sp"
                 android:gravity="center"
                 android:textStyle="bold"
-
                 />
 
             <antonkozyriatskyi.circularprogressindicator.CircularProgressIndicator
@@ -45,7 +44,7 @@
                 android:layout_height="130dp"
                 android:layout_gravity="center_horizontal"
                 android:layout_marginTop="50dp"
-                app:progressColor="@color/dark_blue"
+                app:progressColor="@color/white"
                 app:progressBackgroundColor="@color/grey"
                 app:progressStrokeWidth="18dp"
                 app:direction="clockwise"
@@ -65,7 +64,7 @@
                     android:text="@string/data"
                     android:textStyle="bold"
                     android:textSize="18sp"
-                    android:textColor="@color/dark_blue"
+                    android:textColor="@color/white"
                     />
 
                 <TextView
@@ -74,7 +73,7 @@
                     android:layout_height="wrap_content"
                     android:layout_alignParentRight="true"
                     android:text="0.0 MB"
-                    android:textColor="@color/dark_blue"
+                    android:textColor="@color/white"
                     android:textStyle="bold"
                     android:textSize="18sp"
                     />
@@ -93,7 +92,7 @@
                     android:id="@+id/label"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:textColor="@color/dark_blue"
+                    android:textColor="@color/white"
                     android:text="@string/wi_fi"
                     android:textStyle="bold"
                     android:textSize="18sp"
@@ -105,7 +104,7 @@
                     android:layout_height="wrap_content"
                     android:layout_alignParentRight="true"
                     android:text="0.0 MB"
-                    android:textColor="@color/dark_blue"
+                    android:textColor="@color/white"
                     android:textStyle="bold"
                     android:textSize="18sp"
                     />
@@ -118,7 +117,6 @@
                 android:layout_marginBottom="15dp"
                 />
         </LinearLayout>
-
 
     </LinearLayout>
